CampusX/12_YoutubeChatbot/pyproject.toml
[project]
name = "12-youtubechatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-groq>=1.1.1",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "youtube-transcript-api>=1.2.3",
]
